+++ __brick__/lib/core/constants/app_constants.dart
@@ -1,4 +1,0 @@
-class AppConstants {
-  static const String appName = '{{project_name.titleCase()}}';
-  static const int pageSize = 20;
-}