+++ __brick__/lib/core/injectable/modules/generic.dart
@@ -1,0 +1,8 @@
+import 'package:injectable/injectable.dart';
+import 'package:logger/logger.dart';
+
+@module
+abstract class GenericModule {
+  @lazySingleton
+  Logger getLogger() => Logger();
+}